# api/archive/openapi.v1.1.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MealMap API",
    "version": "1.1.1",
    "summary": "Meal planning, grocery computation, collaboration, and optional AI helpers.",
    "description": "Versioned REST API with WebSocket/SSE realtime channels. User-scoped ingredients, global seeded categories, planner with configurable slots (breakfast, snackAM, lunch, snackPM, dinner), grocery computation with simple unit conversion, N-trip splitting, templates, exports, notifications, sharing, and optional features (nutrition, import, optimizer, normalization, substitutions, barcode, OCR).",
    "termsOfService": "https://api.mealmap.app/terms",
    "contact": { "name": "MealMap API Support", "email": "support@mealmap.app" },
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "https://api.mealmap.app/v1", "description": "Production" },
    { "url": "https://staging.api.mealmap.app/v1", "description": "Staging" }
  ],
...
  "security": [{ "bearerAuth": [], "tenant": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token (expires ~15m). Use refresh to obtain new access."
      },
      "tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "Optional client identifier for rate partitioning."
      }
    },
    "parameters": {
      "pageLimit": {
        "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 },
        "description": "Items per page (default 25, max 100)."
      },
      "pageCursor": {
        "name": "cursor", "in": "query", "schema": { "type": "string" },
        "description": "Opaque cursor for pagination."
      },
...
    },
...
    "responses": {
...
      "NoContent": { "description": "No content." }
    },
    "schemas": {
...
      "UUID": { "type": "string", "format": "uuid" },
      "ISODateTime": { "type": "string", "format": "date-time" },
      "Unit": {
        "type": "string",
        "enum": ["g", "kg", "ml", "l", "piece", "pack"]
      },
...
      "Quantity": {
        "type": "object",
        "properties": {
          "amount": { "type": "number", "minimum": 0 },
          "unit": { "$ref": "#/components/schemas/Unit" }
        },
        "required": ["amount", "unit"]
      },
      "RecipeItem": {
        "type": "object",
        "properties": {
          "ingredientId": { "$ref": "#/components/schemas/UUID" },
          "quantity": { "$ref": "#/components/schemas/Quantity" },
          "packageNote": { "type": "string", "description": "Optional free-text packaging note" }
        },
        "required": ["ingredientId", "quantity"]
      },
      "Recipe": {
        "type": "object",
        "description": "A reusable set of ingredient items with an optional external URL.",
        "properties": {
          "id": { "$ref": "#/components/schemas/UUID" },
          "name": { "type": "string" },
          "externalUrl": { "type": "string", "format": "uri", "nullable": true },
          "items": {
            "type": "array", "maxItems": 150,
            "items": { "$ref": "#/components/schemas/RecipeItem" }
          },
          "createdAt": { "$ref": "#/components/schemas/ISODateTime" },
          "updatedAt": { "$ref": "#/components/schemas/ISODateTime" }
        },
        "required": ["id", "name", "items", "createdAt"]
      },
      "PlannerItem": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/UUID" },
          "date": { "type": "string", "format": "date" },
          "slot": { "type": "string", "enum": ["breakfast", "snackAM", "lunch", "snackPM", "dinner"] },
          "recipeId": { "$ref": "#/components/schemas/UUID", "nullable": true },
          "adHocItems": { "type": "array", "items": { "$ref": "#/components/schemas/RecipeItem" } },
          "portions": { "type": "number", "minimum": 0.1, "default": 1.0 },
          "addedByUserId": { "$ref": "#/components/schemas/UUID" }
        },
        "required": ["id", "date", "slot", "portions"]
      },
...
      "Template": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/UUID" },
          "name": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/PlannerItem" } },
          "createdAt": { "$ref": "#/components/schemas/ISODateTime" }
        },
        "required": ["id", "name", "items", "createdAt"]
      },
...
    }
  },
  "paths": {
...
    "/recipes/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/UUID" } }],
      "get": { "tags": ["Recipes"], "summary": "Get recipe", "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Recipe" } } } } }, "security": [{ "bearerAuth": [] }] },
      "patch": { "tags": ["Recipes"], "summary": "Update recipe", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Recipe" } } } }, "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Recipe" } } } } }, "security": [{ "bearerAuth": [] }] },
      "delete": { "tags": ["Recipes"], "summary": "Delete recipe", "responses": { "204": { "$ref": "#/components/responses/NoContent" } }, "security": [{ "bearerAuth": [] }] }
    },
...
    "/templates": {
      "get": { "tags": ["Templates"], "summary": "List templates", "parameters": [{ "$ref": "#/components/parameters/pageLimit" }, { "$ref": "#/components/parameters/pageCursor" }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Template" } }, "nextCursor": { "type": "string", "nullable": true } } } } } } }, "security": [{ "bearerAuth": [] }] },
      "post": { "tags": ["Templates"], "summary": "Create template", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Template" } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Template" } } } } }, "security": [{ "bearerAuth": [] }] }
    },
...
  },
...
}